recall original raytracing module & test new learn.raytracing functions
