+++ odak/raytracing/boundary.py
@@ -1,7 +1,6 @@
 import numpy as np
-import torch
 from ..tools.vector import distance_between_two_points, closest_point_to_a_ray
-from ..raytracing.primitives import is_it_on_triangle, is_it_on_triangle_batch, center_of_triangle, sphere_function, cylinder_function
+from ..raytracing.primitives import is_it_on_triangle, center_of_triangle, sphere_function, cylinder_function
 from ..raytracing.ray import create_ray_from_two_points, propagate_a_ray
 
 
@@ -63,7 +62,7 @@ def intersect_w_surface(ray, points):
     normal = get_triangle_normal(points)
     if len(ray.shape) == 2:
         ray = ray.reshape((1, 2, 3))
-    if len(points.shape) == 2:
+    if len(points) == 2:
         points = points.reshape((1, 3, 3))
     if len(normal.shape) == 2:
         normal = normal.reshape((1, 2, 3))
@@ -80,49 +79,6 @@ def intersect_w_surface(ray, points):
         distance = distance.reshape((distance.shape[1]))
     return normal, distance
 
-def intersect_w_surface_batch(ray, triangle):
-    """
-    Parameters
-    ----------
-    ray          : torch.tensor
-                   A vector/ray (2 x 3). It can also be a list of rays (n x 2 x 3).
-    points       : torch.tensor
-                   Set of points in X,Y and Z to define a planar surface. It can also be a list of triangles (m x 3 x 3).
-
-    Returns
-    ----------
-    normal       : torch.tensor
-                   Surface normal at the point of intersection (m x n x 3 x 3).
-    distance     : torch.tensor
-                   Distance in between starting point of a ray with it's intersection with a planar surface (m x n).
-    """
-    normal = get_triangle_normal(triangle)
-    if len(ray.shape) == 2:
-        ray = ray.unsqueeze(0)
-    if len(triangle.shape) == 2:
-        triangle = triangle.unsqueeze(0)
-    if len(normal.shape) == 2:
-        normal = normal.unsqueeze(0)
-
-    f = normal[:, None, 0] - ray[None, :, 0]
-    distance = (torch.bmm(normal[:, None, 1], f.permute(0, 2, 1)).squeeze(1) / torch.mm(normal[:, 1], ray[:, 1].T)).T
-
-    new_normal = torch.zeros((triangle.shape[0], )+ray.shape)
-    new_normal[:, :, 0] = ray[None, :, 0] + (distance[:, :, None] * ray[:, None, 1]).permute(1, 0, 2)
-    new_normal[:, :, 1] = normal[:, None, 1]
-    new_normal = torch.nan_to_num(
-                                  new_normal,
-                                  nan = float('nan'),
-                                  posinf = float('nan'),
-                                  neginf = float('nan')
-                                 )
-    distance = torch.nan_to_num(
-                                distance,
-                                nan = float('nan'),
-                                posinf = float('nan'),
-                                neginf = float('nan')
-                               )
-    return new_normal, distance.T
 
 def get_triangle_normal(triangle, triangle_center=None):
     """
@@ -190,10 +146,10 @@ def intersect_w_triangle(ray, triangle):
 
     Parameters
     ----------
-    ray          : torch.tensor
-                   A vector/ray (2 x 3). It can also be a list of rays (n x 2 x 3).
-    points       : torch.tensor
-                   Set of points in X,Y and Z to define a planar surface. It can also be a list of triangles (m x 3 x 3).
+    ray          : ndarray
+                   A vector/ray.
+    triangle     : ndarray
+                   Set of points in X,Y and Z to define a single triangle.
 
     Returns
     ----------
@@ -207,63 +163,6 @@ def intersect_w_triangle(ray, triangle):
         return 0, 0
     return normal, distance
 
-def intersect_w_triangle_batch(ray, triangle):
-    """
-    Definition to find intersection points of rays with triangles. Returns False for each variable if the rays doesn't intersect with given triangles.
-
-    Parameters
-    ----------
-    ray          : torch.tensor
-                   vectors/rays (n x 2 x 3).
-    triangle     : torch.tensor
-                   Set of points in X,Y and Z to define triangles (m x 3 x 3).
-    triangle     : torch.tensor
-                   Set of points in X,Y and Z to define triangles (m x 3 x 3).
-
-    Returns
-    ----------
-    normal          : torch.tensor
-                      Surface normal at the point of intersection (m x n x 2 x 3).
-    distance        : List
-                      List of distances (k, k <= n*m) in between starting point of a ray with it's intersection with a planar surface.
-    intersect_ray   : List
-                      List of intersecting rays (k x 2 x 3, k <= n).
-    intersect_normal: List
-                      List of intersecting normals (k x 2 x 3, k <= n*m).
-    check           : torch.tensor
-                      Boolean tensor (m x n) indicating whether each ray intersects with a triangle or not.
-    """
-    if len(triangle.shape) == 2:
-       triangle = triangle.unsqueeze(0)
-    if len(ray.shape) == 2:
-       ray = ray.unsqueeze(0)
-
-    normal, distance = intersect_w_surface_batch(ray, triangle)
-
-    check = is_it_on_triangle_batch(normal[:, :, 0], triangle)
-
-    flat_check = check.flatten()
-    flat_normal = normal.view(-1, normal.size(-2), normal.size(-1))
-    flat_ray = ray.repeat(normal.size(0), 1, 1)
-    flat_distance = distance.flatten()
-
-    filtered_normal = torch.masked_select(flat_normal, flat_check.unsqueeze(-1).unsqueeze(-1).repeat(1, 2, 3))
-    filtered_ray = torch.masked_select(flat_ray, flat_check.unsqueeze(-1).unsqueeze(-1).repeat(1, 2, 3))
-    filtered_distnace = torch.masked_select(flat_distance, flat_check)
-
-    check_count = check.sum(dim=1).tolist()
-    split_size_ray_and_normal = [count*2*3 for count in check_count]
-    split_size_distance = [count for count in check_count]
-
-    normal_grouped = torch.split(filtered_normal, split_size_ray_and_normal)
-    ray_grouped = torch.split(filtered_ray, split_size_ray_and_normal)
-    distance_grouped = torch.split(filtered_distnace, split_size_distance)
-
-    intersecting_normal = [g.view(-1, 2, 3) for g in normal_grouped if g.numel() > 0]
-    intersecting_ray = [g.view(-1, 2, 3) for g in ray_grouped if g.numel() > 0]
-    new_distance = [g for g in distance_grouped if g.numel() > 0]
-
-    return normal, new_distance, intersecting_ray, intersecting_normal, check
 
 def get_sphere_normal(point, sphere):
     """
@@ -470,5 +369,3 @@ def intersect_w_sphere(ray, sphere):
         get_sphere_normal
     )
     return normal, distance
-
-
