+++ test/test_learn_ray_create_ray_from_grid_w_luminous_angle.py
@@ -5,7 +5,7 @@ import odak
 import torch
 
 def test():
-    rays_from_grid_w_luminous_angle = odak.raytracing.create_ray_from_grid_w_luminous_angle(
+    rays_from_grid_w_luminous_angle = odak.learn.raytracing.create_ray_from_grid_w_luminous_angle(
         center = torch.tensor([0, 0, 0]),
         size = [2, 2],
         no = [2, 2],
